WIP asyncChangeTriggers. Had my head up my ass when I wrote the tests for this apparently.
I got confused building this it seems, the test and implementation are a mess, will have to slightly refactor the test so the fucking implementation does what it's supposed to. In the future the design documents will have to be cleaner in order for the implementation to be smoother I guess?